13.12.19 9:09

+++ tests/test.cpp
@@ -30,13 +30,13 @@ EXPECT_EQ(std::any_cast<std::string>(object["lastname"]),
 "Ivanov");
 EXPECT_EQ(std::any_cast<bool>(object["islegal"]),
 false);
-EXPECT_EQ(std::any_cast<int>(object["age"]),
+EXPECT_EQ(std::any_cast<double>(object["age"]),
 25);
 
 auto marks = std::any_cast<Json>(object["marks"]);
-EXPECT_EQ(std::any_cast<int>(marks[0]),
+EXPECT_EQ(std::any_cast<double>(marks[0]),
 4);
-EXPECT_EQ(std::any_cast<int>(marks[1]),
+EXPECT_EQ(std::any_cast<double>(marks[1]),
 5);
 
 auto address = std::any_cast<Json>(object["address"]);
